Split into components, WIP backend
