+++ postgresql/vite-postgres/src/App.css
@@ -15,6 +15,7 @@
 .title {
   display: flex;
   width: 100vw;
+  margin-bottom: 25px;
 
   justify-content: center;
   align-items: center;
@@ -30,6 +31,8 @@
 
   width: auto;
   height: auto;
+  margin-left: 10px;
+  margin-right: 10px;
 }
 
 .header-1 {
@@ -41,7 +44,8 @@
   display: flex;
   flex-direction: column;
 
-  width: 100vw;
+  width: 100%;
+  gap: 10px;
 }
 
 .form-inputs {
@@ -69,4 +73,9 @@
 
   width: 250px;
   gap: 5px;
+}
+
+.form-submit input {
+  margin-left: 200px;
+  width: 250px;
 }
